pyproject.toml
 [project]
name = "triosdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.35.0",
    "ipykernel",
    "passlib",
    "asteval>=1.0.6",
    "invoke>=2.2.0",
    "pandas>=2.3.2",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.1",
]
